packages/analytics/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stereo-spot-analytics"
version = "0.1.0"
description = "Gather conversion metrics for StereoSpot (AWS CloudWatch, future GCP)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "stereo-spot-shared",
    "stereo-spot-aws-adapters",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
]

[project.scripts]
stereo-spot-analytics = "stereo_spot_analytics.__main__:main"

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.1.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
